Rename IgnorableSerializerContractResolver to ExtendedSerializerContractResolver
